+++ cypress-shared/integration/b2b/2.10-approver_scenarios_orgA_costcenter2.spec.js
@@ -22,7 +22,7 @@ import {
   quoteShouldbeVisibleTestCase,
   quoteShouldNotBeVisibleTestCase,
   useQuoteForPlacingTheOrder,
-  verifySubTotal,
+  verifyTotal,
 } from '../../support/b2b/quotes.js'
 
 describe('Organization A - Cost Center A2 - Approver Scenario', () => {
@@ -58,7 +58,7 @@ describe('Organization A - Cost Center A2 - Approver Scenario', () => {
     quotes.Approver2.updateQuote,
     role.Approver2.dropDownText
   )
-  verifySubTotal(quotes.Approver2.updateQuote)
+  verifyTotal(quotes.Approver2.updateQuote)
   fillContactInfo()
   verifyAddress(costCenter2.addresses)
   verifyPayment(false)